Move session hooks to module, sessions is now extendable and distributable. We can move it to it's own service
